update[deploy]: 进一步完善deploy。 1. 目前deploy支持dep阶段的代码部署 2. 目前deploy不支持prod阶段的代码部署。

+++ deploy/bin/reset.sh
@@ -6,16 +6,23 @@
 # 3.调用deploy.sh启动服务
 # 注意：由于1和2的原因，请仅在开发测试阶段使用本脚本！
 
-
 # 重置数据库
 # i. 请在`XXXXXX`处设置相应的root密码
 # ii. 同时请注意root密码放在脚本是非常危险的，因此这里仅仅是用于开发测试阶段。
+ROOT=root
+PASSWORD=
+
+if test -z "$PASSWORD"
+then
+  echo "请设置云主机MySQL的root账号密码"
+  exit -1
+fi
+
 cd /home/ubuntu/deploy/litemall-db
 cat litemall_schema.sql > db.sql
 cat litemall.sql >> db.sql
-mysql -h localhost -u root -pXXXXXX < db.sql
+mysql -h localhost -u $ROOT -p$PASSWORD < db.sql
 rm db.sql
-cd /home/ubuntu/deploy/litemall-db
 
 # 删除storage文件夹内文件
 cd /home/ubuntu/deploy/litemall-os-api/storage
